// src/shared/types/cv-graphql.ts
export enum UserRole {
  Employee = "Employee",
  Admin = "Admin",
}
// ...
export interface Project {
  id: string;
  created_at: string;
  name: string;
  internal_name: string;
  domain: string;
  start_date: string;
  end_date?: Nullable<string>;
  description: string;
  environment: string[];
}
// ...
type Nullable<T> = T | null;